fixed some warnings
1. **`sam/reader.rs:39`** — collapsed nested `if let` + `if` into single `if let && &&`
2. **`record_store.rs:350`** — removed redundant `as u32` cast (already u32)
3. **`region_buf.rs:579`** — `vec![0u64]` → `[0u64]` (array, not heap-allocated vec)
4. **`cram/slice.rs:880`** — push-after-creation → `vec![...]` initializer

+++ crates/seqair/src/bam/record_store.rs
@@ -347,7 +347,7 @@ mod tests {
         // Write cigar: M ops
         let cigar_start = var_start + name_len as usize;
         for i in 0..n_cigar_ops as usize {
-            let op = (seq_len as u32) << 4; // M op
+            let op = seq_len << 4; // M op
             let off = cigar_start + i * 4;
             if off + 4 <= raw.len() {
                 raw[off..off + 4].copy_from_slice(&op.to_le_bytes());
